Fix crash when parsing duplicate verses separated by a paragraph marker
